add rudimentary healthcheck

+++ entrypoint/10-patch-opvn-configs.sh
@@ -19,21 +19,25 @@ log::error() {
 }
 
 main() {
-    local config_dir="${OPENVPN_CONFIG_DIR:-/config/client}"
-    local pass_file="${OPENVPN_PASS_FILE:-/config/credentials}"
+    local config_dir="${CONFIG_DIR:-/config/client}"
+    local pass_file="${PASS_FILE:-/config/credentials}"
 
     sed -i "s:auth-user-pass:auth-user-pass ${pass_file}:" "${config_dir}"/*.ovpn
 
-    >"$pass_file"
-    chmod 600 "$pass_file"
-
     if [[ -z "$USERNAME" ]]; then
-        log::error "Environment variable 'USERNAME' is unset."
+        log::warn "Environment variable 'USERNAME' is unset."
     fi
 
     if [[ -z "$PASSWORD" ]]; then
         log::warn "Environment variable 'PASSWORD' is unset. Defaulting to 'hunter2'."
     fi
+
+    if [[ -f "$pass_file" ]]; then # A pass file was bind mounted in
+        return
+    fi
+
+    >"$pass_file"
+    chmod 600 "$pass_file"
 
     printf '%s\n' "$USERNAME" > "$pass_file"
     printf '%s\n' "${PASSWORD:-hunter2}" >> "$pass_file"
